refactor: Delegate signup dialog view responsibility to SignupDialogViewer (#222)
